FEM: ccx tools, move helper into fem utils module
